Use matrix transformations to draw the shimmer instead of rotating a drawn rect
Summary: Instead of trying to compute the correct size of the shimmer draw rect by an incorrect heuristic, we instead apply a matrix transformation to the shader. This means we dont have to calculate a specially sized draw rect to ensure we cover the entire view, but it does mean that we need to enlongate the translation distance by the opposite dimension of the right triangle that is formed with the shimer's tilt and the non primary axis.

Reviewed By: sjkirby

Differential Revision: D8729182

fbshipit-source-id: eb8d948a9006a6193713113013fb8d82a94af3a8