Add show + hide functionality (#86)
Summary:
PR for https://github.com/facebook/shimmer-android/issues/85

It doesn't seem like the best implementation since I'm just nulling the Shimmer but it does the job.

I've tested this in the sample app by depending on the library project rather than the snapshot and added another button for toggling it on and off but haven't included any of that in this PR. If you would like me to I'll update it.
Pull Request resolved: https://github.com/facebook/shimmer-android/pull/86

Differential Revision: D16053297

Pulled By: xiphirx

fbshipit-source-id: 9f50e1f56c1f2272fbb2da70dbea3680da6e03f1